+++ variables.tf
@@ -22,7 +22,7 @@ variable "storage_type" {
 }
 
 variable "storage_throughput" {
-  description = "Storage throughput value for the DB instance. This setting applies only to the `gp3` storage type."
+  description = "Storage throughput value for the DB instance. See `notes` for limitations regarding this variable for `gp3`"
   type        = number
   default     = null
 }
@@ -34,7 +34,7 @@ variable "storage_encrypted" {
 }
 
 variable "kms_key_id" {
-  description = "The ARN for the KMS encryption key. If creating an encrypted replica, set this to the destination KMS ARN. If storage_encrypted is set to true and kms_key_id is not specified the default KMS key created in your account will be used"
+  description = "The ARN for the KMS encryption key. If creating an encrypted replica, set this to the destination KMS ARN. If storage_encrypted is set to true and kms_key_id is not specified the default KMS key created in your account will be used. Be sure to use the full ARN, not a key alias."
   type        = string
   default     = null
 }
@@ -164,9 +164,9 @@ variable "multi_az" {
 }
 
 variable "iops" {
-  description = "The amount of provisioned IOPS. Setting this implies a storage_type of 'io1'"
+  description = "The amount of provisioned IOPS. Setting this implies a storage_type of 'io1' or `gp3`. See `notes` for limitations regarding this variable for `gp3`"
   type        = number
-  default     = 0
+  default     = null
 }
 
 variable "publicly_accessible" {